Revamp prompting and logging
