+++ code/core/src/common/js-package-manager/NPMProxy.ts
@@ -2,7 +2,7 @@ import { existsSync, readFileSync } from 'node:fs';
 import { platform } from 'node:os';
 import { join } from 'node:path';
 
-import { logger } from 'storybook/internal/node-logger';
+import { prompt } from 'storybook/internal/node-logger';
 import { FindPackageVersionsError } from 'storybook/internal/server-errors';
 
 import { findUpSync } from 'find-up';
@@ -172,7 +172,7 @@ export class NPMProxy extends JsPackageManager {
 
         return this.mapDependencies(parsedOutput, pattern);
       } catch (err) {
-        logger.warn(`An issue occurred while trying to find dependencies metadata using npm.`);
+        prompt.warn(`An issue occurred while trying to find dependencies metadata using npm.`);
         return undefined;
       }
     }
@@ -192,6 +192,7 @@ export class NPMProxy extends JsPackageManager {
       command: 'npm',
       args: ['install', ...this.getInstallArgs()],
       cwd: this.cwd,
+      stdio: prompt.getPreferredStdio(),
       ignoreError: true,
     });
   }
@@ -218,7 +219,7 @@ export class NPMProxy extends JsPackageManager {
     return this.executeCommand({
       command: 'npm',
       args: ['install', ...args, ...this.getInstallArgs()],
-      stdio: 'pipe',
+      stdio: prompt.getPreferredStdio(),
       cwd: this.primaryPackageJson.operationDir,
     });
   }
