--- code/core/src/node-logger/prompts.ts
@@ -1,34 +1,0 @@
-import {
-  debug,
-  error,
-  getLogLevel,
-  intro,
-  log,
-  logBox,
-  outro,
-  setLogLevel,
-  step,
-  warn,
-} from './logger';
-import * as clackFunctions from './prompt-functions';
-import * as promptFunctions from './prompt-functions-prompts';
-import { executeTask, executeTaskWithSpinner } from './tasks';
-
-const USE_CLACK = process.env.USE_CLACK === 'false' ? false : true;
-const functions = USE_CLACK ? clackFunctions : promptFunctions;
-
-export const prompt = {
-  ...functions,
-  intro,
-  outro,
-  logBox,
-  log,
-  warn,
-  error,
-  step,
-  debug,
-  executeTask,
-  executeTaskWithSpinner,
-  setLogLevel,
-  getLogLevel,
-};
