+++ code/core/src/node-logger/prompts/prompt-provider-clack.ts
@@ -1,0 +1,66 @@
+import * as clack from '@clack/prompts';
+
+import { logTracker } from './log-tracker';
+import type {
+  ConfirmPromptOptions,
+  MultiSelectPromptOptions,
+  PromptOptions,
+  SelectPromptOptions,
+  SpinnerInstance,
+  TaskLogInstance,
+  TaskLogOptions,
+  TextPromptOptions,
+} from './prompt-provider-base';
+import { PromptProvider } from './prompt-provider-base';
+
+export class ClackPromptProvider extends PromptProvider {
+  private handleCancel(result: unknown | symbol, promptOptions?: PromptOptions) {
+    if (clack.isCancel(result)) {
+      if (promptOptions?.onCancel) {
+        promptOptions.onCancel();
+      } else {
+        clack.cancel('Operation canceled.');
+        process.exit(0);
+      }
+    }
+  }
+
+  async text(options: TextPromptOptions, promptOptions?: PromptOptions): Promise<string> {
+    const result = await clack.text(options);
+    this.handleCancel(result, promptOptions);
+    logTracker.addLog('prompt', options.message, { choice: result });
+    return result.toString();
+  }
+
+  async confirm(options: ConfirmPromptOptions, promptOptions?: PromptOptions): Promise<boolean> {
+    const result = await clack.confirm(options);
+    this.handleCancel(result, promptOptions);
+    logTracker.addLog('prompt', options.message, { choice: result });
+    return Boolean(result);
+  }
+
+  async select<T>(options: SelectPromptOptions<T>, promptOptions?: PromptOptions): Promise<T> {
+    const result = await clack.select<T>(options);
+    this.handleCancel(result, promptOptions);
+    logTracker.addLog('prompt', options.message, { choice: result });
+    return result as T;
+  }
+
+  async multiselect<T>(
+    options: MultiSelectPromptOptions<T>,
+    promptOptions?: PromptOptions
+  ): Promise<T[]> {
+    const result = await clack.multiselect<T>(options);
+    this.handleCancel(result, promptOptions);
+    logTracker.addLog('prompt', options.message, { choice: result });
+    return result as T[];
+  }
+
+  spinner(): SpinnerInstance {
+    return clack.spinner();
+  }
+
+  taskLog(options: TaskLogOptions): TaskLogInstance {
+    return clack.taskLog(options);
+  }
+}
